requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.0
google-auth==2.34.0
google-api-python-client==2.147.0
pydantic==2.9.2
